+++ Exception/RouteStructureException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace MarfaTech\Bundle\RabbitQueueBundle\Exception;
+
+class RouteStructureException extends RabbitQueueException
+{
+}