+++ Producer/RabbitMqProducerInterface.php
@@ -6,5 +6,5 @@ namespace MarfaTech\Bundle\RabbitQueueBundle\Producer;
 
 interface RabbitMqProducerInterface
 {
-    public function put(string $queueName, $data, array $options = []);
+    public function put(string $queueName, $data, array $options = [], string $routingKey = '');
 }