+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+### Changed
+- Refactoring namespace to `MarfaTech`.
 
 ## [2.1.2] - 2021-07-15
 ### Changed