+++ CHANGELOG.md
@@ -19,8 +19,8 @@
 - [BC BREAK] Renamed parameter `wakeapp_rabbit_queue.connection.heartbeat` to `marfatech_rabbit_queue.connection.heartbeat`.
 - [BC BREAK] Renamed parameter `wakeapp_rabbit_queue.consumer.idle_timeout` to `marfatech_rabbit_queue.consumer.idle_timeout`.
 - [BC BREAK] Renamed parameter `wakeapp_rabbit_queue.consumer.wait_timeout` to `marfatech_rabbit_queue.consumer.wait_timeout`.
-- [BC BREAK] Change interface `RabbitmqProducerInterface::put` added parameter `routingKey` and `properties`.
-- [BC BREAK] Change interface `PublisherInterface::publish` added parameter `routingKey` and `properties`.
+- [BC BREAK] Changed interface `RabbitmqProducerInterface::put` added parameter `routingKey` and `properties`.
+- [BC BREAK] Changed interface `PublisherInterface::publish` added parameter `routingKey` and `properties`.
 
 ## [2.1.2] - 2021-07-15
 ### Changed